Add Markdown to HTML conversion
Wiki pages are now presented as html instead of markdown.
Markdown conversion is done without use of external libraries.

Markdown is also santised before conversion, to ensure injection
attacks are not possible by writing html/javascript in the markdown.
This allows us to use the |safe tag when loading the page contents so
the tags we support are rendered as proper html.